+++ packages/netlify-cms-widget-markdown/src/MarkdownControl/plugins/SelectAll.js
@@ -1,0 +1,14 @@
+import isHotkey from 'is-hotkey';
+
+const SelectAll = () => ({
+  onKeyDown(event, editor, next) {
+    const isModA = isHotkey('mod+a', event);
+    if (!isModA) {
+      return next();
+    }
+    event.preventDefault();
+    return editor.moveToRangeOfDocument();
+  },
+});
+
+export default SelectAll;
