+++ packages/netlify-cms-widget-markdown/src/types.js
@@ -1,0 +1,6 @@
+export const SLATE_DEFAULT_BLOCK_TYPE = 'paragraph';
+
+export const SLATE_BLOCK_PARENT_TYPES = [
+  'list-item',
+  'quote',
+];
